requirement.txt
attrs==23.2.0
Automat==22.10.0
backoff==2.2.1
beautifulsoup4==4.12.3
black==23.12.1
brotlipy==0.7.0
bump2version==1.0.1
bumpversion==0.6.0
Cerberus==1.3.5
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.5
cssselect==1.2.0
decorator==5.1.1
docutils==0.20.1
extruct==0.16.0
filelock==3.13.3
html-text==0.5.2
html5lib==1.1
hyperlink==21.0.0
idna==3.6
importlib_metadata==7.1.0
incremental==22.10.0
iniconfig==2.0.0
isodate==0.6.1
isort==5.13.2
itemadapter==0.8.0
itemloaders==1.1.0
jaraco.classes==3.4.0
jaraco.context==4.3.0
jaraco.functools==4.0.0
jmespath==1.0.1
jstyleson==0.0.2
keyring==25.0.0
loguru==0.7.2
lxml==5.2.0
Mako==1.3.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mf2py==2.0.1
more-itertools==10.2.0
msgpack==1.0.8
mypy-extensions==1.0.0
nh3==0.2.17
packaging==24.0
parsel==1.9.0
pathspec==0.12.1
pdoc3==0.10.0
pkginfo==1.10.0
platformdirs==4.2.0
pluggy==1.4.0
Protego==0.3.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDispatcher==2.0.7
Pygments==2.17.2
pyOpenSSL==24.1.0
pyparsing==3.1.2
pyRdfa3==3.6.2
pytest==7.4.4
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
queuelib==1.6.2
rdflib==7.0.0
readme_renderer==43.0
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
ruff==0.0.269
scrapfly-sdk==0.8.15
Scrapy==2.11.1
sentry-sdk==1.44.0
service-identity==24.1.0
six==1.16.0
soupsieve==2.5
tldextract==5.1.2
twine==5.0.0
Twisted==24.3.0
typing_extensions==4.10.0
urllib3==2.2.1
w3lib==2.1.2
webencodings==0.5.1
zipp==3.18.1
zope.interface==6.2
